refactor(failed-messages): process new messages, delete successful ones, retry failed ones
modularize the code a bit as it was growing to be confusing with the layout we had. this helps us handle the different ways to handle messages as well. we dont want to delete from PEL as thats the messages redis hasnt seen yet. we should process those. we can delete them entirely from the stream if they are successful. we ack the ones that dont get a 201 from SCRAM. this lets us then go back and loop the acked messages and retry, without impacting the new PEL messages